+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `Tuple2-8` specializations of left/right product rotation
 - `CheckedEffect`, an `Effect` variant that can throw checked exceptions 
 - `CheckedFn1#checked`, convenience static factory method to aid inference 
+- `LiftA3-8`, higher-arity analogs to `LiftA2` 
 
 ### Deprecated
 - `AddAll` semigroup, in favor of the monoid that no longer mutates any argument